+++ internal/component/database_observability/postgres/collector/connection_info.go
@@ -5,6 +5,7 @@ import (
 	"regexp"
 	"strings"
 
+	"github.com/grafana/alloy/internal/util"
 	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/atomic"
 )
@@ -36,7 +37,7 @@ func NewConnectionInfo(args ConnectionInfoArguments) (*ConnectionInfo, error) {
 		Help:      "Information about the connection",
 	}, []string{"provider_name", "provider_region", "db_instance_identifier", "engine"})
 
-	args.Registry.MustRegister(infoMetric)
+	util.MustRegisterOrGet(args.Registry, infoMetric)
 
 	return &ConnectionInfo{
 		DSN:        args.DSN,