+++ internal/runtime/module_test.go
@@ -243,10 +243,12 @@ func TestDuplicateIDList(t *testing.T) {
 		return len(nc.ModuleIDs()) == 1
 	}, 5*time.Second, 100*time.Millisecond)
 
-	// This should panic with duplicate registration.
-	require.PanicsWithError(t, "duplicate metrics collector registration attempted", func() {
-		_, _ = nc.NewModule("t1", nil)
-	})
+	// This should no longer panic with duplicate registration, but still be rejected when Run is called.
+
+	mod1Dup, err := nc.NewModule("t1", nil)
+	require.NoError(t, err)
+	err = mod1Dup.Run(ctx)
+	require.Error(t, err)
 }
 
 func testModuleControllerOptions(t *testing.T) *moduleControllerOptions {