+++ internal/static/integrations/v2/apache_http/apache_http.go
@@ -11,6 +11,7 @@ import (
 	integrations_v2 "github.com/grafana/alloy/internal/static/integrations/v2"
 	"github.com/grafana/alloy/internal/static/integrations/v2/common"
 	"github.com/grafana/alloy/internal/static/integrations/v2/metricsutils"
+	"github.com/grafana/alloy/internal/util"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
@@ -93,7 +94,7 @@ func (ah *apacheHandler) createHandler() (http.HandlerFunc, error) {
 
 	return func(w http.ResponseWriter, r *http.Request) {
 		registry := prometheus.NewRegistry()
-		registry.MustRegister(aeExporter)
+		util.MustRegisterOrGet(registry, aeExporter)
 		h := promhttp.HandlerFor(registry, promhttp.HandlerOpts{})
 		h.ServeHTTP(w, r)
 	}, nil